Updated UI back to normal
