+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "dotenv.enableAutocloaking": false
+  "dotenv.enableAutocloaking": true
 }
