+++ city-ranker/components/auth/logout.jsx
@@ -11,7 +11,7 @@ export default function LogOut({ onClick }) {
                 <button className='loginButton' style={{ color: 'black', fontWeight: 'bold', fontSize: '10px' }}>
                     LOGOUT
                 </button>
-                <img style={{ cursor: 'pointer', width: '24px' }} src='/assets/logout.svg' />
+                <img style={{ cursor: 'pointer', width: '24px' }} src='/logout.svg' />
             </div>
         </a>
     )
