+++ city-ranker/components/main/cityRow/ScoresContainer.jsx
@@ -6,17 +6,17 @@ export default function ScoresContainer({ walkScore, bikeScore, transitScore, ci
         <>
             <div className='scoresContainer'>
                 <div className='scoreContainer'>
-                    <img alt={'Walking Icon for walk score'.concat(cityID)} src='/assets/walk.svg' />
+                    <img alt={'Walking Icon for walk score'.concat(cityID)} src='/walk.svg' />
                     <Pie percentage={walkScore} colour="blue" />
                 </div>
 
                 <div className='scoreContainer'>
-                    <img alt={'Biking Icon for bike score'.concat(cityID)} src='/assets/bike.svg' />
+                    <img alt={'Biking Icon for bike score'.concat(cityID)} src='/bike.svg' />
                     <Pie percentage={bikeScore} colour="blue" />
                 </div>
 
                 <div className='scoreContainer'>
-                    <img alt={'Train Icon for transit score'.concat(cityID)} src='/assets/train.svg' />
+                    <img alt={'Train Icon for transit score'.concat(cityID)} src='/train.svg' />
                     <Pie percentage={transitScore} colour="blue" />
                 </div>
             </div>
