+++ city-ranker/pages/api/[address].js
@@ -1,8 +1,3 @@
-var express = require('express');
-var router = express.Router();
-const fetch = require('node-fetch');
-require('dotenv').config();
-
 export default function handler(request, response) {
     response.status(200).json({
       body: request.body,
@@ -12,16 +7,16 @@ export default function handler(request, response) {
   }
   
 
-router.get('/:address', function (req, res, next) {
+// router.get('/:address', function (req, res, next) {
 
-    let url = `https://maps.googleapis.com/maps/api/geocode/json?address=${req.params.address}&key=${process.env.GOOGLE}`;
+//     let url = `https://maps.googleapis.com/maps/api/geocode/json?address=${req.params.address}&key=${process.env.GOOGLE}`;
 
-    let options = { method: 'GET' };
+//     let options = { method: 'GET' };
 
-    fetch(url, options)
-        .then(res => res.json())
-        .then(json => res.send(json))
-        .catch(err => console.error('error:' + err));
-});
+//     fetch(url, options)
+//         .then(res => res.json())
+//         .then(json => res.send(json))
+//         .catch(err => console.error('error:' + err));
+// });
 
-module.exports = router;
+// module.exports = router;
