--- city-ranker/pages/api/index.js
@@ -1,5 +1,0 @@
-import { NowRequest, NowResponse } from '@vercel/node'
-
-export default (req, res) => {
-  return res.json({ message: 'Hello World' })
-}
